requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
aiosqlite>=0.19.0
httpx>=0.25.0
websockets>=12.0
cryptography>=41.0.0
python-multipart>=0.0.6
structlog>=23.2.0
python-dotenv>=1.0.0
